service-data-metadata/requirements.txt
Pillow==9.3
confluent_kafka